backend/utils/requirements.txt
Flask
flask-cors
pdfminer.six
PyPDF2
pytesseract
Pillow
pandas
